Fixed the compile errors.
Should have compiled and uploaded it before pushing the changes. Sorry.

Hopefully should work? Doesn't print anything to PC serial anymore and
uses the Arduino's hardware serial to communicate with the SIM808.
You'll need to swap the jumpers on the SIM808 accordingly.

Should send updates to my home PC server with the settings as is. I
already took care of the port forwarding and listen server for them.