Last commit/update of semester
Additional tweaks to code prior to C-day. Hardware interrupts to detect
on/off status mostly work, powering the GPS board off while the
equipment is off needs some adjustment, and the SD cache is write-only
right now.

The newest versions use the Arduino Mega instead of the Arduino Uno and
connects to the GPS shield through Serial1 instead of through software
serial on pins 7 and 8, as previous.